atlas-antenna-demo/requirements.txt
pynmea2>=1.18.0
pyserial>=3.5
